requirements.txt
annotated-types==0.7.0
anyio==4.12.1
apscheduler==3.10.4
pytz==2025.2
bcrypt==5.0.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
mysql-connector-python==9.6.0
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-dotenv==1.2.2
python-multipart==0.0.22
razorpay==2.0.1
requests==2.32.5
starlette==0.52.1
typing_extensions==4.15.0
uvicorn==0.42.0
fpdf2==2.8.7
Pillow==12.1.1
